\name{Cho}
\alias{Cho}
\docType{data}
\title{
expression matrix and clinical information from Cho et al
}
\description{
expression matrix and clinical information associated with response to anti-PD-1 therapy in non-small cell lung cancer
}
\usage{data("Cho")}
\references{
Cho J. W., Hong M. H., Ha S. J., et al. Genome-wide identification of differentially methylated promoters and enhancers associated with response to anti-PD-1 therapy in non-small cell lung cancer. Exp Mol Med. Sep 2020;52(9):1550-1563.
}
\examples{
data(Cho)
}
\keyword{datasets}
