\name{CRMA}
\alias{CRMA}
\docType{data}
\title{
Anti-CTLA4 resistance MAGE genes (CRMA)
}
\description{
Anti-CTLA4 resistance MAGE genes (CRMA), including MAGEA3, CSAG3, CSAG2, MAGEA2, MAGEA2B, CSAG1, MAGEA12 and MAGEA6.
}
\usage{data("CRMA")}
\references{
Shukla S. A., Bachireddy P., Schilling B., et al. Cancer-Germline Antigen Expression Discriminates Clinical Outcome to CTLA-4 Blockade. Cell. Apr 19 2018;173(3):624-633.e628.
}
\examples{
data(CRMA)
}
\keyword{datasets}
