\name{Gide}
\alias{Gide}
\docType{data}
\title{
expression matrix and clinical information from Gide et al
}
\description{
expression matrix and clinical information associated with response to Anti-PD-1 Monotherapy and Anti-PD-1/Anti-CTLA-4 Combined Therapy in melanoma
}
\usage{data("Gide")}
\details{
In the clinical information, Alias=='EDT' means on-treatment collected samples; Alias=='PRE' means pretreatment collected samples
}
\references{
Gide T. N., Quek C., Menzies A. M., et al. Distinct Immune Cell Populations Define Response to Anti-PD-1 Monotherapy and Anti-PD-1/Anti-CTLA-4 Combined Therapy. Cancer Cell. Feb 11 2019;35(2):238-255.e236.
}
\examples{
data(Gide)
}
\keyword{datasets}
