\name{house}
\alias{house}
\docType{data}
\title{
house keeping genes 
}
\description{
house keeping genes for ICIResponseRes
}
\usage{data("house")}
\references{
Sommariva S., Tarricone R., Lazzeri M., et al. Prognostic Value of the Cell Cycle Progression Score in Patients with Prostate Cancer: A Systematic Review and Meta-analysis. Eur Urol. Jan 2016;69(1):107-115.
}
\examples{
data(house)
}
\keyword{datasets}
