\name{INFG}
\alias{INFG}
\docType{data}
\title{
IFN gamma as a response biomarker by Ayers et al.
}
\description{
IFNgamma was founded as a response biomarker by Ayers et al.
}
\usage{data("INFG")}
\references{
Ayers M., Lunceford J., Nebozhyn M., et al. IFN-related mRNA profile predicts clinical response to PD-1 blockade. J Clin Invest. Aug 1 2017;127(8):2930-2940.
}
\examples{
data(INFG)
}
\keyword{datasets}
