\name{IPSG}
\alias{IPSG}
\docType{data}
\title{
gene set in Immunophenoscore
}
\description{
Immunophenoscore (IPS) was calculated by a script supplied by The Cancer Immunome Atlas
}
\usage{data("IPSG")}
\references{
Charoentong P., Finotello F., Angelova M., et al. Pan-cancer Immunogenomic Analyses Reveal Genotype-Immunophenotype Relationships and Predictors of Response to Checkpoint Blockade. Cell Rep. Jan 3 2017;18(1):248-262.
}
\examples{
data(IPSG)
}
\keyword{datasets}
