\name{Jung}
\alias{Jung}
\docType{data}
\title{
expression matrix and clinical information from Jung et al
}
\description{
expression matrix and clinical information associated with response to anti-PD-1 therapy in in lung cancer and melanoma
}
\usage{data("Jung")}
\references{
Jung H., Kim H. S., Kim J. Y., et al. DNA methylation loss promotes immune evasion of tumours with high mutation and copy number load. Nat Commun. Sep 19 2019;10(1):4278.
}
\examples{
data(Jung)
}
\keyword{datasets}
