\name{Kim}
\alias{Kim}
\docType{data}
\title{
expression matrix and clinical information from Kim et al
}
\description{
expression matrix and clinical information associated with response to anti-PD-1 therapy in metastatic gastric cancer
}
\usage{data("Kim")}
\references{
Kim S. T., Cristescu R., Bass A. J., et al. Comprehensive molecular characterization of clinical responses to PD-1 inhibition in metastatic gastric cancer. Nat Med. Sep 2018;24(9):1449-1458.
}
\examples{
data(Kim)
}
\keyword{datasets}
