\name{PD1highCD8Tscore-package}
\alias{PD1highCD8Tscore-package}
\alias{PD1highCD8Tscore}
\docType{package}
\title{
PD1highCD8Tscore
}
\description{
A transcriptional signature to quantify the fraction of PD-1hiCD8+ T cells in tumor environment. 
}
\details{
Patients with high signature scores are more likely to response to immune checkpoint inhibitors, and are associated with increased clinical outcome.
getSigScore, calculate PD-1hi CD8+ T cell signature score 
getIPS, Immunophenoscore (IPS) was calculated by a script supplied by The Cancer Immunome Atlas
ICIResponse, explore the association between a marker and ICI therapy in datasets.
Six datasets about expression matrix and clinical information associated with response to immune checkpoint inhibitors
}
\author{
Zhenyu Yang
Maintainer: Yulan Deng
}
\references{
Tumor-infiltrating PD-1hiCD8+ T cell as potential biomarker for immune checkpoint inhibitors in multiple cancers: an integrative study.
}
\keyword{ package }
