\name{Riaz}
\alias{Riaz}
\docType{data}
\title{
expression matrix and clinical information from Riaz et al
}
\description{
expression matrix and clinical information associated with response to anti-PD-1 therapy in melanoma
}
\usage{data("Riaz")}
\details{
In the clinical information, time=='On' means on-treatment collected samples; time=='Pre' means pretreatment collected samples
}
\references{
Riaz N., Havel J. J., Makarov V., et al. Tumor and Microenvironment Evolution during Immunotherapy with Nivolumab. Cell. Nov 2 2017;171(4):934-949.e916.
}
\examples{
data(Riaz)
}
\keyword{datasets}
