\name{TLS}
\alias{TLS}
\docType{data}
\title{
gene sets for Tertiary lymphoid structures
}
\description{
gene sets for Tertiary lymphoid structures from Cabrita et al
}
\usage{data("TLS")}
\references{
Cabrita R., Lauss M., Sanna A., et al. Tertiary lymphoid structures improve immunotherapy and survival in melanoma. Nature. Jan 2020;577(7791):561-565.
}
\examples{
data(TLS)
}
\keyword{datasets}
